utility/requirements.txt
click==8.1.3
Flask==2.2.2
grpcio==1.50.0
grpcio-tools==1.50.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.9
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
